msm: mdss: Wait for VSYNC after continuous splash handoff
After the pipes are unstaged during continuous splash
handoff the driver should wait for VSYNC to make sure
the changes are flushed before moving on. Currently it
sleeps for 20ms but this might not be enough especially
on low fps configurations causing pipe hang. Changing
that to wait for VSYNC interrupt instead.

Change-Id: Ic722a7d6c39c36084f35f55e7e899476e34f607f
Signed-off-by: Rajkumar Subbiah <rsubbia@codeaurora.org>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>
Signed-off-by: DarkDampSquib <andrin.geiger1998@gmail.com>