blkdev: Refactoring block io latency histogram codes
The current io_latency_state structure includes entries for read
and write requests. There are special types of write commands
such as sync and discard (trim) commands, and the current
implementation is not general enough if we want to separate
latency histogram for such special commands.

This change makes io_latency_state structure request-type neutral.
It also changes to print the latency average.

Change-Id: Iaa4fbcc428110741b69ed42c02db234e2e2dd5c7
Signed-off-by: Hyojun Kim <hyojun@google.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: EcrosoftXiao <xy19990903@gmail.com>