BACKPORT: kasan, arm64: clean up KASAN_SHADOW_SCALE_SHIFT usage
Upstream commit 9597e74396c02c8189215dcdfc08088a9b8d8087.

This is a follow up patch to the series I sent recently that cleans up
KASAN_SHADOW_SCALE_SHIFT usage (which value was hardcoded and scattered
all over the code). This fixes the one place that I forgot to fix.

The change is purely aesthetical, instead of hardcoding the value for
KASAN_SHADOW_SCALE_SHIFT in arch/arm64/Makefile, an appropriate variable
is declared and used.

Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Change-Id: I63e3c8578b42727bf3ffe135de6888d0f7344022
Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Bug: 128674696
Signed-off-by: DarkDampSquib <andrin.geiger1998@gmail.com>