binfmt_elf: Don't allocate memory dynamically in load_elf_binary
The dynamic memory allocations in load_elf_binary can be replaced by
large stack allocations that bring the total frame size for the function
up to about 4 KiB. We have 16 KiB of stack space, and this function does
not have deep function call depths, so such a large stack allocation is
safe. This improves performance by eliminating the overhead of dynamic
memory allocation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[fixed for newer 4.14 tags]
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>

Change-Id: Ifa714857992d64cbf1e466e4bc3e9b08e291d8cb
Signed-off-by: DarkDampSquib <andrin.geiger1998@gmail.com>