mm: util: Guarentee memory access to applications
In an ideal world, Android would request only the amount of memory it
needs, and nothing more or less. Unfortunately, some apps request too
little memory, requiring the kernel to guess how much (if any) memory to
overcommit.

We could skip this process by setting overcommit_memory to 1, so
applications can request up to 50% of the system memory. We will rely on
lmkd to garbage collect unfreed memory.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: DarkDampSquib <andrin.geiger1998@gmail.com>